# docs/swagger.yaml
definitions:
  auth_handlers.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  auth_handlers.loginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: P@ssword123
        type: string
    required:
    - email
    - password
    type: object
  auth_handlers.registerRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: P@ssword123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  chat_handlers.ActiveUsersResponse:
    properties:
      active_users:
        example:
        - user_1
        - user_2
        items:
          type: string
        type: array
    type: object
  chat_handlers.CreateChatRequest:
    properties:
      receiver_id:
        type: string
      sender_id:
        type: string
    required:
    - receiver_id
    - sender_id
    type: object
  chat_handlers.CreateChatResponse:
    properties:
      chat_id:
        type: string
      chat_name:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: Backend API for Flutter Chat App.
  title: Go Chat App API
  version: "1.0"
paths:
  /api/v1/auth/activate/{link}:
    get:
      consumes:
      - text/plain
      description: Activation of account after registration step
      parameters:
      - description: Generated link's end side edge
        in: path
        name: link
        required: true
        type: string
      produces:
      - text/html
      responses:
        "201":
          description: Returns the html page of success
          schema:
            type: string
        "400":
          description: 'Could not get the link from url path. Possible "error" values
            : [BAD_REQUEST]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "410":
          description: 'Activation link is expired . Possible "error" values : [ACTIVATION_TIME_EXPIRED]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: 'Server failed to process . Possible "error" values : [INTERNAL_SERVER_ERROR]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Activate
      tags:
      - Registration
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Basic Login operation with email and password
      parameters:
      - description: Provide your creds for creation new session on certain device
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth_handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Session was successfully created
          schema:
            $ref: '#/definitions/auth_handlers.AuthResponse'
        "400":
          description: 'Possible "error" values: [EMPTY_AUTH_CREDS] '
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: "Possible \"error\" values: [INVALID_CREDS, \"any\tother\tprintable\terrors\"]"
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: 'Server failed to process . Possible "error" values : [INTERNAL_SERVER_ERROR]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Login
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      consumes:
      - text/plain
      description: Logging out by revoking session
      parameters:
      - description: Insert 'Bearer <AccessToken>'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "204":
          description: Logged out successfully
        "401":
          description: 'Possible "error" values: [EMPTY_AUTH_CREDS]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: 'Server failed to process . Possible "error" values : [INTERNAL_SERVER_ERROR]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Log out
      tags:
      - Authorization
  /api/v1/auth/me:
    get:
      consumes:
      - text/plain
      description: Endpoint for checking authorization
      parameters:
      - description: Insert 'Bearer <your_token>'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'user_id": some_user_id (int)'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 'Possible "error" values: [INTERNAL_SERVER_ERROR]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Checking authorization
      tags:
      - Profile
  /api/v1/auth/refresh:
    post:
      consumes:
      - text/plain
      description: Get new pair of token
      parameters:
      - description: Insert 'Bearer <RefreshToken>'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: New refreshed token pair returned. So update them both
          schema:
            $ref: '#/definitions/auth_handlers.AuthResponse'
        "401":
          description: "Possible \"error\" values: [INVALID_CREDS,\"all\tother\terrors\"]"
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: 'Server failed to process . Possible "error" values : [INTERNAL_SERVER_ERROR]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: refresh
      tags:
      - Authorization
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: For initiating registration process with sending activation link
      parameters:
      - description: User registration details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/auth_handlers.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Activation link is successfully sent
        "400":
          description: "Request is invalid. Possible \"error\" values:[USER_ALREADY_EXISTS,\"all\tothers\tare\tcaused\tby\tthe\trequest's\tincorrectness\"]"
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: 'Server failed to process. Possible "error" values : [INTERNAL_SERVER_ERROR]'
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Register
      tags:
      - Registration
  /api/v1/chats/create:
    post:
      consumes:
      - application/json
      description: Initiates a new chat between two users or returns existing. If
        a 1-to-1 chat already exists between these users, it returns the existing
        chat details.
      parameters:
      - description: Insert 'Bearer <AccessToken>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create Chat Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/chat_handlers.CreateChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created or retrieved chat
          schema:
            $ref: '#/definitions/chat_handlers.CreateChatResponse'
        "400":
          description: Invalid request body or missing fields
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Create or get a 1-to-1 chat
      tags:
      - chats
  /api/v1/users/active:
    get:
      description: Returns an array of all currently connected User IDs.
      parameters:
      - description: Insert 'Bearer <AccessToken>'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat_handlers.ActiveUsersResponse'
      summary: List active users
      tags:
      - users
  /ws:
    get:
      description: Initiates a WebSocket connection. Requires user_id in query.
      parameters:
      - description: Insert 'Bearer <AccessToken>'
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: query
        name: user_id
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
      summary: Upgrade to WebSocket
      tags:
      - websocket
swagger: "2.0"